Move tabular tests from test_core to test_tabular
